requirements.txt
alembic==0.9.6
astroid==1.5.3
blinker==1.4
cffi==1.11.2
click==6.7
colorama==0.3.9
cryptography==1.7.2
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
Flask-Uploads==0.2.1
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
psycopg2==2.7.3.2
pyasn1==0.3.7
pycparser==2.18
PyJWT==1.5.3
pylint==1.7.4
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.14
Werkzeug==0.12.2
wrapt==1.10.11
